Fixed the lat/long problem, and CSV problem.
CSV no longer crashes when there are not variable headers in the “data”
tab. Fixed parser to look for “longitude” and “latitude” in the
metadata worksheet header. It no longer looks for a long specific
latitude string, etc. New problem in the Holocene folder when trying to
parse the data columns.