Type Error from name_to_jsonld
Problem was that when the template cells of “Data” are not present,
then the cell_down_datasheet would not know where to stop looping. When
floats and ints passed into name_to_jsonld it broke the parser. Now, we
look for “Missing Value” or “Data” in the cell value, which is more
reliable.