No longer pass floats through name converter
Problem was that the parser was reading in the variables, and then not
finding the ‘Data’ cell (because it was removed from the template) and
it didn’t know when to stop. It looped over the float data entries, and
they are not meant to be passed through the name converter.