Fix Line Continuation bug. Starting GeoJson
1. Line continuation was having an error where it wouldn’t know what
the key was for the previous loop (or section) and wouldn’t know where
to insert into the dictionary. Now uses an ‘old_key’ to track what the
previous key was.
2. Need to reconfigure, and create a new way to parse the geo block to
adhere to geojson standards. Wrote out the steps of how to do it, and
what it will look like. Now going to start coding it.