Adapted NOAA to LMR test files. GeoJson updates
Not all files have a ‘missing values’ line, so I adapted the data
section to use the “# Data: ” line as a marker instead, because that is
a consistent line in all files. Added more keys to the GeoJson
properties block.

This version successfully parsed all the LMR test files, with the
exception of a few files that had Unicode errors because they are not
utf-8 format.