Scrapped MongoDB+Django+Python setup
This combination is not playing together well at all and seems to be
very experimental. Many of the projects that are trying to make MongoDB
and Django work are old and only support Python 2.7 or earlier, or work
with Django 1.5 and earlier. Tutorials and documentation that exist are
incomplete and won’t complete without many errors.