DOI Resolver, capable of multiple pubs
The DOI resolver has been split into the Integrated DOI, and the Standalone DOI. For testing, we can use the standalone resolver with takes in json data. All doi id's are now compiled to a list before running. This is because it is possible to have multiple doi's for the same file. The output pub is now a list of dictionaries. Each doi (or pub) now appends its own output (dict) to the final pub list.