Updated Schema and Upload
Schema has been updated with given definitions. Anything without a
definition says to reference external resource. Upload has major
changes. Lots of completion on the form and the map and coordinates are
working pretty well. Working to add option to drag markers and making
the coordinate data transferrable to UserData