Upload + Show content working
When a user uploads text or json files, the files are uploaded to the
tmp folder and then immediately retrieved. The contents are displayed
on the page. This does not work with excel files because they are
encoded and can’t be simply displayed. TO DO: Need to spawn a child
process and run the uploaded file through the python parser before
returning contents back to the page.