"Contact Us" Form working with Nodemailer
The contact us form on the bottom of the index page is now functional
and currently sends to my personal gmail account. Can be modified to a
dedicated LiPD email in the future. I used Nodemailer and an angular
controller to handle the post data and transport the email.
Additional: Removed all unused modules and move apps, modules, and js
into separate folders for convenience.