Parser: Bagit DOI regex and integration
Final touches on compiling the proper regex to catch all DOIs parsed,
and to add it to the process flow.
- Regex for recognizing DOIs and non-DOI entries
- Proper handling for empty strings, manual citations, noaa citations,
and other links that are incorrectly placed in the DOI field.
- Creates new fields and places data in correct locations when
improperly placed in the DOI field.
- Quarantine.txt logs file names and info about error. Lists files that
have complications fetching data with DOI id, don’t have a DOI id, or
have unknown info in the DOI field.
- Creates a field that marks whether a file was created manually or
data was fetched from doi.org. (PubDataUrl)
- Process will loop and fetch data for each individual DOI id found.
Each DOI id returns one dictionary of its own data which is appended to
the final pub list.