Overall script working. Needs refinement.
The major functionality of the converter is working. There are a few
things that need to be added and completed.
Changes:
- Class is fully documented, and easily reusable.
- All sections are written to the text file successfully
- txt file is created and built in the tmp folder. It is not moved into
the final output folder until the file successfully closes
- If there is an error opening a zip flie, the tmp folder is deleted
(Previously the tmp folders were piling up when the script hit errors
since tmp is normally deleted at the end)
Needs Work:
- Need example files with “Missing Value” fields, so I know how to
capture/write that value when necessary.
- Need naming conventions for CSV files so numeric data can be added to
output noaa text files. (Chron, Data, etc). I can get CSV data, but not
sure what to do with it.
-