Fix: Data formatting problem
Fix: Data was not being formatted correctly (regarding ‘\n’ missing).
This was because when the string was longer than 15 digits, and being
truncated to fit into 15 digit precision, the new line character (at
the end of every CSV file line) was being cut off. Therefore, no new
line was made. I fixed this by not truncating the last item in every
line.
Alternate: If this becomes a problem later, the last number on each
line can be truncated, checked for a new line character, and add one if
it’s missing.