BibJSON and Bug
- Formatted the publication data to match the BibJSON standard
- Bug: The Resolver runs for each publication dictionary. However, each
time the resolver ran, it would also append the fetched data as a new
entry, thus infinitely extending the loop and causing a hang. Fixed the
resolver to have it replace the existing data with the fetched data
in-place.
Next push: Track bug that causes XLRD not to open 3 test files.