Pub_on marker fix
I added a new way to handle multiple publications in the last push. It
required making a marker that turned on when I hit the publication
section. However, I never put in a step to turn off the marker when I
hit the end of the publication section, so it was adding extra empty
strings to all subsequent metadata. I added in the step to turn off the
marker and all is well now.