Funding Fix. Removing empty values.
- Funding was filling in empty cells with “” to make things easier.
However, since we want to eliminate empty entries, compile_fund()
needed to be reworked. It now iterates agency and grant at the same
time and keeps adding cells until both are blank. If agency is full and
grant is not, it will create an entry without the grant cell. (or vice
versa).
- Added a global list for EMPTY strings. I’ll be using global lists for
similar functions in the future. Currently there are many of the same
exact lists being created and destroyed throughout the script. It’s
more cluttered.
- No longer adds Climate Interpretation if it is empty.