Remove empties from Publication
At the moment, the only section that still had empty entries in it was
the Publication Dictionaries. Everything we do invokes the DOI Resolver
to get the most updated publication information. Therefore, the best
spot to remove any entries from the publication data was right before
the DOI Resolver finishes. Any empty entries at root level will be
removed, and any identifiers that do not contain a DOI ID will be
removed as well.