Bug: Single values in geo coordinates
When a value was present, it passed that float value to the function.
When a value was not present, it passed an empty list to the function.
The function got confused because it was receiving two different types
and didn't know what to do with an empty list. Fixed it be using None
as a placeholder, and handling the None when inside the function.