TimeSeries: Additions
- Documentation: Function name changes for consistency. Trying to keep
LiPD_Library and TimeSeries_Library commands similar so that it’s easy
on the user. All user functions are CamelCase. All background functions
(that the user shouldn’t be using) are underscored.
-  Exception handling: In case the user decides accidentally puts in
bad commands, don’t break the program.
- Sorting: When showing LiPD and TSO filenames, put them in
alphabetical order so you don’t have to hunt so much.
- LiPD object loading: Pre-screen the incoming LiPD data before setting
the data. Remove empty info so functions later on can assume data is
good.
- Convert Object->Geo: previous exception handling was omitting geo
coordinates altogether when calling extractTimeSeries. Fixed so geo is
added normally now.
- Convert Object->Geo->meanElev: some datasets are trying to put
meanElev in the “properties” section of data. Added another step to
check for that and make sure to cast the value to float if that happens.
- Convert() lpd_name: most dataset’s filename and ‘dataSetName’ field
are the same name. However, some are slightly different and that was
causing problems converting TimeSeries back to LiPD. Decided to attach
the LiPD filename (more consistent) to each TimeSeries object to avoid
mismatches. No longer use ‘dataSetName’.
- JSONS->doi: Added function that specifically checks for and removes
empty DOIs and ‘identifier’ dictionaries.
- JSONS -> json+csv: Added functions that will take a file with
combined JSON metadata and column values, split them, and return
separate CSV only and JSON only dictionaries.