LiPD.net - new approach
- Switched to a “function over fashion” uploader. They other uploader
looked a bit prettier, but the page length and size grew pretty fast
with files that contained many tables. This caused a lot of unneeded
scrolling and trying to figure out where you were on the page.  Rather
than try to figure out a way to condense or create some type of
accordion tree hierarchy, I found a module that let me create an
interactive, editable JSON tree that folds at each level and is much
simpler. It allows a whole lipd metadata file to appear in the window,
and expand as you choose.