LiPD.net: Map location, validation, editing
- Validation will check uploaded data against LiPD 1.2 structure and
deem it valid or invalid.
- All invalid errors ( incorrect data types for a given field,
incorrect structure, missing required data fields) will two-way bind
with the error box on the page showing an error count and the cause of
the errors
- All less serious errors will show as “warnings” in a warning box in
the side panel. This might include unknown keys found in the file, or
unable to find helpful data (like coordinates) to show location on the
map, etc.
- Validate button: As before, data fields can be edited, deleted, or
added. However, rather than put expensive $watch operations on the page
to have continuous validation, I put a button that will clear warnings,
errors, and  re-evaluate all the metadata. This is in case you make
changes and want to make sure the file is still valid.