Large LiPD upload file bug
ODP1098.lpd and other large files were taking much longer to upload. I
thought it was because it had more data to process, but it turns out
that the validator was attempting to validate hidden Mac OS files and
getting hung up. I’m now checking for “._” at the beginning of
filenames, which is the prefix of a hidden system file.