LiPD.net: ES6 -> ES5
- Decreased font size and increased the input boxes on the json tree.
In safari, it was doing some scaling and making the text inside the
json tree unreadable. Found a good middle ground.
- I was having some trouble with JS hoisting variables above the  block
scope of functions, which was causing errors. I decided to use “let”
declarations which works in chrome, but not in safari. I’ve decided to
“transpile” the ES6 code into ES5 because it’s much easier and less
time consuming at this point.
- Fixed Lat Lon issue that was causing markers not to be placed on the
map
- Added modals for viewing file content. Data is not yet linked to
modals.
- Added “missingValue” to the list of required fields.