Validator: Bagit working
Added the “gladstone” module for processing Bagit in Node.js. It’s not
a popular module and not kept up, so I had to make a fair amount of
changes and fixes to make it work properly and make it structured the
way we want it.

Process: Create a random generated tmp folder/string. Do all the file
creation, bagging, and archiving during POST. Send back the tmp string
to client. Client does GET with tmp string in params. Use tmp string to
locate lipd file and initiate download.