Cleaning out ngValidate controller
The validate controller has a lot going on. It’s much more unruly that
it should be. I’ve split up a lot of the functions into external
scripts that can be called. This will be helpful on the LiPD Creation
page because now all of these functions are available to any page,
rather than only the validator. Also, the controller is less painful to
look at now.