Lipd.net: Create-from-scratch ready for prime time
- Update to angular 1.4.8. Anything past that, and it was causing
errors with the google maps module.
- Updated angular material css to v1.1.4
- Updated angular material js to v1.1.4
- Added a “fixer” for malformed or incorrect publication “Author” data
- Added dropdown for Geo Country field. ISO 3166 standard list.
- Added dropdown for archiveType (root and column level). Using
predefined list from Excel template.
- Added checkboxes for adding extra properties to each data column.
Allows for entering your own property if it’s not in the predefined
list from Excel template.
- Fixed checking for required fields in publications.
- Fixed checking for required fields in columns
- Fixed checking for required fields in table.
- Changed all indexing on the client-side to 0-indexed to be consistent
with the data structure in the back-end. It was becoming confusing when
tracking errors when client-side was 1-indexed and back-end was
0-indexed.
- Separated the create.jade and validator.jade into modular pieces that
are easier to work on and easier to track. Each section has its own
jade file.
- Added CSV parser. Choose from 4 delimiter options, choose if a header
row is present, and all data is parsed. Function creates necessary
table columns on the page after parsing.
- Enabled fields that were previously disabled, because they were
required fields that were unable to be edited if a file was uploaded
without that data. There was no way to enter data to get a valid file.