Modals
Modals are not mostly working. Removed some old code and modules and
stuck to angular bootstrap ui. When you click on a filename on the left
hand side, it’ll open up a large modal that shows the contents of the
file in a “pre” tag. i.e. you can view the bagit md5’s, other bagit
data, csv data, and jsonld in a pretty printed format (in case the json
you want to see isn’t a field listed on the page.

The CSV is not formatted in a nice way, so next is to get a table of
CSV data to show in the modal.