Web: Create page for v1.3
- Updated a lot of the create page to be v1.3 compatible. Paleo and
chron measurement tables now show up for creation / editing as before.
- Since we are no longer attempting to show multiple versions, v1.3 is
the assumed version. If a file was updated from a lesser version, a
message box will pop up and note as such.
- Added warning at the top of the page that ensemble, distribution, and
summary tables are not yet supported.
- Added lipdVersion field back into the view. This has become a more
important field for the user to see these days.
- If lipdVersion is unknown when uploaded, we assume v1.3. We’ll
attempt our best guess work, but a message warns the user that our
attempt may be futile without knowing the version.