Web: ensemble, summary, method, and more
The online creation tool only lets you add/view/edit the first entry of
any chronData/paleoData, and the first entry of that model.

This is shown:
paleoData[0]model[0]summary[any].

This is NOT shown :
paleoData[1]model[0]summary[0]
paleoData[0]model[1]summary[0]

- Added support ensemble tables
- Added support for summary tables
- Added support for method table
- Removed ‘distribution’ from the dropdown list of tables allowed to be
added (this is too complex for now and not worth the hassle. large data)
- Split the jade files for measurement, model, chron, paleo. The views
were getting immense and tough to edit.