Web: Create interpretation now works
How: Click the interpretation checkbox option within a column, and this
opens a modal box with checkboxes for options that you do or don’t want
in your interpretation. There are required , standard, and optional
boxes. Clicking a button creates the interpretation and adds it to the
LiPD data

To Do: Currently does not show the fields for editing, which isn’t
helpful.