Web: Wiki and NOAA validations
- There were some issues with the way wiki validates. For example, it
was attempting to validate even when the “wiki ready” switch was OFF. I
added an “options” object that sends through to the validator. The
options consist of metadata about the page, so I know how to validate.
- NOAA: The validation for the NOAA rules is in effect (when the switch
is turned ON).
- NOAA rules have been updated to the tooltips on the switch
- Partially done with NOAA input fields. Column fields still need to be
added, but root and geo items are visible when the switch is turned ON.