Web: bug fixes and better table ui
- Added expandable and collapsable columns that show the variableName
and column number where available.
- Fixed but where you couldn’t download a file while the completed
captcha was still active. Made it so the clicking download works
whenever successful captcha is active, and triggers captcha when it is
not active.
- v1.3 filenames: previously we were grabbing the filename and
datasetname from the inner filenames. (i.e.
“heiser.flagstaff.2017/metadata.jsonld”) However, with v1.3 that
doesn’t work because the inner folder is generically named “bag”. So,
the filename and dsn are now being grabbed from the main “filename.lpd”
as it is uploaded.