#9 Fixing the table indexing issue (visual bug)
The indexing and use of the $parent index is identical in multiple
files. However, the bug was because ng-if creates its own scope, which
caused  another layer away from the $parent loop that we wanted to use.
Since one file (paleo_meas, etc) doesn’t use ng-if, and the other
paleo_buttons, etc) _does_ use ng-if, we saw a different indexing.

The solution was to use ng-init at the $parent loop to persist the
parent index down past the ng-if scope.