Don't autofill coordinates and elevation with 0
When validating, empty coordinate fields were pre-filled with 0 to
prevent issues during validation. We don’t want these fields to be 0
because that may give incorrect info since those are valid numbers.
Now, we pre-fill with nulls. It still passes validation and returns the
right errors, and the input fields remain blank.