#24 #25 Changes with column data
- Added an “Edit” button that allows you to delete any fields that are
in the “additional fields” section
- Removed the checkboxes for adding and removing ‘additional fields’.
Now, it is a dropdown box with options, and a button to add the
selected field. This looks much better and makes more sense
programatically.
- Better handling for nested data: calibration, hasResolution,
physicalSample, and other block data can now be added and edited. They
have block buttons and modal windows just like the interpretation data.
These are a work in progress because I don’t know what fields should be
“common field” options yet, but there is a custom field that allows the
user to enter fields and data.