#24 : Better interaction for nested column data
- Modals are more universal now. I made the modal controller and
functions for nested column data work for arrays and object data. That
means data like interpretation and calibration will pass through the
same process (which was formerly two processes).
- I left these modals modular so they can expand as column data
(possibly) expands  in the future
- Trying to add or remove restricted / automated fields from the column
data will produce a warning alert and not allow the action to complete
(i.e. TSid)
- Other minor changes, css and other.