#16 fix and modular tooltips
- Wiki and NOAA fields are added to new columns and existing columns
when the respective switches are turned on
- Tooltips are modular now. Given the amount of fields that are on the
page, it was getting pretty hairy to track down tooltips and make
changes. Also, dynamic fields, like those in the “additional fields”
section of columns, were not able to have tooltips. Now, they can
because the getTooltip function can be repeated in ng-repeat. All
tooltips are stored in an object in ng_create.js. This leaves open the
option in the future of fetching and compiling fields and tips from the
Wiki and so forth.