#27 : Fixed the Adding columns bug
Instead of keeping a watcher on the dataSetName field and changing the
CSV filenames (possibly) multiple times in a session, I decided to look
for matches in the filename base when updating json table data and csv
scope data. Since every indexing is unique, (ex:
“paleo0measurement0.csv”), then either that will already exist in the
scope, or not. Look for this match (without the datasetname) and use it
or create new data.