Query page finished
- all fields and checkboxes are now working
- user’s choices are validated against the set of query constraints
before sending out the HTTP request
- if the query isn’t valid, the errors display on the page to show the
user what needs to be fixed to complete the query. Same format as the
playground page, to keep the system consistent and familiar.
- results are shown on the page in a two column display with a direct
download button and link button to the wiki source page for each
respective dataset.
- The original python script sends requests with an array of field name
variations. ex. “marine sediment” and “Marine Sediment” are both part
of the query. All archiveType, inferredVariableType, and
proxyObservationType fields get dynamic variations of the field names
added to the query. Dynamic variations are created based on the main
list of field names using a regex and toUpperCase()