#42 : Added safeguard to file upload
The problem was not as described in the issue. The file in question did not have any json data inside of the metadata.jsonld file. Inside of that file was just a string with the dataset name "BJ8-2D03-2D13GGC.Linsley.2010".

This fix will check after upload if the json data is an object or a string. If it's a string, we'll notify the user that something went wrong and to notify us on the issue tracker because there is something wrong with the file.