#29 : File list updates
The file list wasn't being updated or bound to the changing playground data. It would show the original data from the initial file upload OR if you were creating a file from scratch, it would not show filenames for data that you are creating. Now, the metadata is bound to the main json data behind the page. The csv and bagit data is updated each time the data is validated. There is not a way to fully bind those two types of data.