#53 : wiki direct upload
Did some integration with the wiki direct file upload. It's not working, but the scaffolding and parts are setup. I'm waiting for a response from Varun to get more info on how to do this as the API docs don't answer my questions. Right now, I attempt to direct upload a valid wiki LiPD file, and it gets a 200 response, but I don't see the file on the wiki site and not sure what the issue is.