#56 Mismatch columns
I stopped trying to look for a fast solution, and went for a slow / working solution. Instead of checking the first row of data to see how many columns are in a data table, I now look at ALL rows, and choose the longest row, to determine the number of columns. This gets around the problem of empty cells and counting the wrong amount of columns.