#60  Send Lipd to Lipdverse
Users can now create a LiPD file and send it to Lipdverse. It first goes to our LiPD dropbox for review, and then is moved to Lipdverse once approved.

This does not include an e-mail system to notify LiPD Managers of when a new file is uploaded and in the queue. That will come soon.