DOI API fix
For unknown reasons, the DOI API stopped working due to CORS restrictions. It previously worked for years with no problem. I moved the API call to the back end so it's done in node, and the response is passed to the front end. The rest of the parsing is the same.